Move Android shared pref behind an interface for easier testing
