+++ app/src/main/java/fi/methics/musap/sdk/internal/async/BindKeyTask.java
@@ -4,11 +4,11 @@ import android.content.Context;
 
 import fi.methics.musap.sdk.api.MusapCallback;
 import fi.methics.musap.sdk.api.MusapException;
-import fi.methics.musap.sdk.extension.MusapSscdInterface;
 import fi.methics.musap.sdk.internal.datatype.MusapKey;
 import fi.methics.musap.sdk.internal.datatype.SscdInfo;
 import fi.methics.musap.sdk.internal.discovery.KeyBindReq;
-import fi.methics.musap.sdk.internal.discovery.MetadataStorage;
+import fi.methics.musap.sdk.internal.discovery.AndroidMetadataStorage;
+import fi.methics.musap.sdk.internal.discovery.SharedPrefStorage;
 import fi.methics.musap.sdk.internal.util.AsyncTaskResult;
 import fi.methics.musap.sdk.internal.util.MLog;
 import fi.methics.musap.sdk.internal.util.MusapAsyncTask;
@@ -30,7 +30,7 @@ public class BindKeyTask extends MusapAsyncTask<MusapKey> {
         try {
             MusapKey key = sscd.bindKey(req);
             MLog.d("BindKeyTask Got MUSAP key");
-            MetadataStorage storage = new MetadataStorage(context.get());
+            AndroidMetadataStorage storage = new AndroidMetadataStorage(new SharedPrefStorage(context.get()));
 
             SscdInfo activeSscd = sscd.getSscdInfo();
             String       sscdId = sscd.getSscdId();
